pulp_distribution: sort repository versions numerically
When creating a distribution without specifying a repository version,
distribution or publication, the latest repository version should be used.
However, there is a text sort in the implementation which should be numerical.
This leads to choosing version 9 over version 10.

This change adds an Ansible filter to perform the sort correctly.

Fixes: #8