pulp_distribution: fix distribution promotion
Promotion of one distribution to the publication of another distribution
currently fails. We see:

    'str object' has no attribute 'pulp_href'

This change fixes the issue by refactoring the complex jinja logic into
ansible filters.

Also fixes an issue when a distribution has a state of absent - we
should not need to specify a repository or base_path.

Fixes: #34
Fixes: #35