feat: change loop approach in `pulp_repository`
The original loop approach would leak sensitive data in the event the
task fails. This new approach does not have this issue as it loops over
`name` and use an `index_var` to access the contents of the items.

This same approach will also be included in other roles contained within
this repository.