Get users and groups from distinct API requests
An API request to retrieve a list of all users and groups is subject
to pagination, which can break assumptions if your object exists
but is not returned in the first page results.

Query each user and group from the API directly, and then parse the
result to distinguish between objects that are present or absent.